reservation message in de reservation gezet

+++ Projce_B_Code/ConfirmationMessage.cs
@@ -1,7 +1,7 @@
 namespace Project_B;
 class ConfirmationMessage
 {
-    public static void ShowConfirmationMessage(Film film, string date, string email, int total_price)
+    public static void ShowConfirmationMessage(Film film, string date, string email, double total_price)
     {
         Console.WriteLine("Uw reservering is succesvol geplaatst.");
         Console.WriteLine($"Film: {film.Name}");
